+++ src/core/ReelRoulette.Core/Abstractions/CoreInterfaces.cs
@@ -8,6 +8,16 @@ public interface IStorageService<T>
     void Save(T value);
 }
 
+public interface IAtomicUpdateStorageService<T> : IStorageService<T>
+{
+    T Update(Func<T, T> update);
+}
+
+public interface IPathResolver
+{
+    string GetPath();
+}
+
 public interface IBackgroundTaskScheduler
 {
     void Queue(string taskName, Action<CancellationToken> work, CancellationToken cancellationToken = default);